+++ src/export_scripts.ts
@@ -1,7 +1,9 @@
+import { getEnvVars } from "./env.ts";
 import { ConfigData } from "./load_config.ts";
+import { EnvironmentVariables } from "./scripts_config.ts";
 import { validateConfigData } from "./validate_config_data.ts";
 import { validateScript } from "./validate_script.ts";
-import { isWindows, makeFileExecutable, OneOrMore, parseEnvFile } from "./util.ts";
+import { isWindows, makeFileExecutable, OneOrMore } from "./util.ts";
 import { normalizeScript } from "./normalize_script.ts";
 import { CompoundCommandItem } from "./command.ts";
 import { log } from "./logger.ts";
@@ -76,20 +78,11 @@ function exportCommands(commands: CompoundCommandItem[]): string {
       }
       const cmd = commands;
       let res = "";
-      if (cmd.env || cmd.env_file) {
-        const env: Record<string, string> = {};
-        if(cmd.env_file) {
-          Object.assign(env, parseEnvFile(cmd.env_file));
-        }
-        if(cmd.env) {
-          Object.assign(env, cmd.env);
-        }
-        const envVars = Object.entries(env);
-        if (envVars.length > 0) {
-          res += envVars
-            .map(([key, val]) => `${key}="${escape(val, '"')}"`)
-            .join(" ") + " ";
-        }
+      const envVars: EnvironmentVariables | undefined = getEnvVars(cmd);
+      if (envVars) {
+        res += Object.entries(envVars)
+          .map(([key, val]) => `${key}="${escape(val, '"')}"`)
+          .join(" ") + " ";
       }
       res += buildCommandString(cmd) + ' "$@"';
       if (doGroup) res = `( ${res} )`;
