+++ src/util.ts
@@ -16,35 +16,3 @@ export function makeFileExecutable(filePath: string) {
     // Windows
   }
 }
-
-const envCache: Record<string, Record<string, string>> = {};
-
-export function parseEnvFile(envFile: string): Record<string, string> {
-  if (envCache[envFile]) {
-    return Object.assign({}, envCache[envFile]);
-  }
-  const buffer: Uint8Array = Deno.readFileSync(envFile);
-  return envCache[envFile] = new TextDecoder()
-    .decode(buffer)
-    .trim()
-    .split(/\n+/g)
-    .map(val => val.trim())
-    .filter(val => val[0] !== "#")
-    .reduce((env: Record<string, string>, line: string) => {
-      const [name, value] = line
-        .replace(/^export\s+/, "")
-        .split("=", 2)
-        .map(val => val.trim());
-      env[name] = stripeQuotes(value);
-      return env;
-    }, {});
-}
-
-function stripeQuotes(value: string) {
-  const first = value[0];
-  const last = value[value.length - 1];
-  if ((first === '"' || first === "'") && first === last) {
-    return value.slice(1, -1);
-  }
-  return value;
-}
